Add DecimalPlacesProperty and change types to int
- Added a new `DecimalPlacesProperty` to the `NumericValidation` class in `NumericValidation.cs` to handle the number of decimal places for numeric validation.
- Changed the type of `MaximumValue` property setter in `NumericValidation.cs` from `double` to `int`.
- Added new methods `GetDecimalPlaces` and `SetDecimalPlaces` in `NumericValidation.cs` to get and set the `DecimalPlacesProperty`.
- Updated the `OnZZZActivateChanged` method in `NumericValidation.cs` to include the `decimalPlaces` variable.
- Changed the type of various `MinPoint` and `MaxPoint` properties from `double` to `int` in `ISettingsService.cs`.
- Updated the `SettingsService` class to reflect the type changes from `double` to `int` for `MinPoint` and `MaxPoint` properties.
- Updated the `SettingsViewModel` class to reflect the type changes from `double` to `int` for `MinPoint` and `MaxPoint` properties.
- Added `behaviors:NumericValidation.DecimalPlaces="0"` to various `Entry` elements in `SettingsPage.xaml` to specify the number of decimal places for numeric validation.

+++ Vm167Box/Services/ISettingsService.cs
@@ -8,51 +8,51 @@ public interface ISettingsService
 
     string Analog1Name { get; set; }
     string Analog1Unit { get; set; }
-    double Analog1MinPoint { get; set; }
+    int Analog1MinPoint { get; set; }
     double Analog1MinValue { get; set; }
-    double Analog1MaxPoint { get; set; }
+    int Analog1MaxPoint { get; set; }
     double Analog1MaxValue { get; set; }
 
     string Analog2Name { get; set; }
     string Analog2Unit { get; set; }
-    double Analog2MinPoint { get; set; }
+    int Analog2MinPoint { get; set; }
     double Analog2MinValue { get; set; }
-    double Analog2MaxPoint { get; set; }
+    int Analog2MaxPoint { get; set; }
     double Analog2MaxValue { get; set; }
 
     string Analog3Name { get; set; }
     string Analog3Unit { get; set; }
-    double Analog3MinPoint { get; set; }
+    int Analog3MinPoint { get; set; }
     double Analog3MinValue { get; set; }
-    double Analog3MaxPoint { get; set; }
+    int Analog3MaxPoint { get; set; }
     double Analog3MaxValue { get; set; }
 
     string Analog4Name { get; set; }
     string Analog4Unit { get; set; }
-    double Analog4MinPoint { get; set; }
+    int Analog4MinPoint { get; set; }
     double Analog4MinValue { get; set; }
-    double Analog4MaxPoint { get; set; }
+    int Analog4MaxPoint { get; set; }
     double Analog4MaxValue { get; set; }
 
     string Analog5Name { get; set; }
     string Analog5Unit { get; set; }
-    double Analog5MinPoint { get; set; }
+    int Analog5MinPoint { get; set; }
     double Analog5MinValue { get; set; }
-    double Analog5MaxPoint { get; set; }
+    int Analog5MaxPoint { get; set; }
     double Analog5MaxValue { get; set; }
 
     string Pwm1Name { get; set; }
     string Pwm1Unit { get; set; }
-    double Pwm1MinPoint { get; set; }
+    int Pwm1MinPoint { get; set; }
     double Pwm1MinValue { get; set; }
-    double Pwm1MaxPoint { get; set; }
+    int Pwm1MaxPoint { get; set; }
     double Pwm1MaxValue { get; set; }
 
     string Pwm2Name { get; set; }
     string Pwm2Unit { get; set; }
-    double Pwm2MinPoint { get; set; }
+    int Pwm2MinPoint { get; set; }
     double Pwm2MinValue { get; set; }
-    double Pwm2MaxPoint { get; set; }
+    int Pwm2MaxPoint { get; set; }
     double Pwm2MaxValue { get; set; }
 
     /// <summary>
